+++ data-transfer-object/src/main/java/org/grnet/status/dtos/profile/aggregation/Group.java
@@ -1,0 +1,35 @@
+package org.grnet.status.dtos.profile.aggregation;
+
+import java.util.List;
+
+public class Group {
+
+    private String name;
+    private String operation;
+    private List<Service> services;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getOperation() {
+        return operation;
+    }
+
+    public void setOperation(String operation) {
+        this.operation = operation;
+    }
+
+    public List<Service> getServices() {
+        return services;
+    }
+
+    public void setServices(List<Service> services) {
+        this.services = services;
+    }
+}
+