+++ data-transfer-object/src/main/java/org/grnet/status/dtos/profile/aggregation/Service.java
@@ -1,0 +1,24 @@
+package org.grnet.status.dtos.profile.aggregation;
+
+public class Service {
+
+    private String name;
+    private String operation;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getOperation() {
+        return operation;
+    }
+
+    public void setOperation(String operation) {
+        this.operation = operation;
+    }
+}
+