+++ data-transfer-object/src/main/java/org/grnet/status/dtos/profile/metric/MetricProfileResponse.java
@@ -1,0 +1,27 @@
+package org.grnet.status.dtos.profile.metric;
+
+import org.grnet.status.dtos.profile.operation.Status;
+
+import java.util.List;
+
+public class MetricProfileResponse {
+
+    private Status status;
+    private List<MetricProfile> data;
+
+    public Status getStatus() {
+        return status;
+    }
+
+    public void setStatus(Status status) {
+        this.status = status;
+    }
+
+    public List<MetricProfile> getData() {
+        return data;
+    }
+
+    public void setData(List<MetricProfile> data) {
+        this.data = data;
+    }
+}