+++ data-transfer-object/src/main/java/org/grnet/status/dtos/profile/metric/ServiceMetrics.java
@@ -1,0 +1,14 @@
+package org.grnet.status.dtos.profile.metric;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+public class ServiceMetrics {
+
+    private String service;
+    private List<String> metrics;
+}