+++ data-transfer-object/src/main/java/org/grnet/status/dtos/profile/operation/Defaults.java
@@ -1,0 +1,32 @@
+package org.grnet.status.dtos.profile.operation;
+
+public class Defaults {
+
+    private String down;
+    private String missing;
+    private String unknown;
+
+    public String getDown() {
+        return down;
+    }
+
+    public void setDown(String down) {
+        this.down = down;
+    }
+
+    public String getMissing() {
+        return missing;
+    }
+
+    public void setMissing(String missing) {
+        this.missing = missing;
+    }
+
+    public String getUnknown() {
+        return unknown;
+    }
+
+    public void setUnknown(String unknown) {
+        this.unknown = unknown;
+    }
+}