+++ data-transfer-object/src/main/java/org/grnet/status/dtos/profile/operation/OperationProfile.java
@@ -1,0 +1,67 @@
+package org.grnet.status.dtos.profile.operation;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.List;
+
+public class OperationProfile {
+
+    private String id;
+    private String date;
+    private String name;
+
+    @JsonProperty("available_states")
+    private List<String> availableStates;
+
+    private Defaults defaults;
+
+    private List<Operation> operations;
+
+    // getters & setters
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getDate() {
+        return date;
+    }
+
+    public void setDate(String date) {
+        this.date = date;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public List<String> getAvailableStates() {
+        return availableStates;
+    }
+
+    public void setAvailableStates(List<String> availableStates) {
+        this.availableStates = availableStates;
+    }
+
+    public Defaults getDefaults() {
+        return defaults;
+    }
+
+    public void setDefaults(Defaults defaults) {
+        this.defaults = defaults;
+    }
+
+    public List<Operation> getOperations() {
+        return operations;
+    }
+
+    public void setOperations(List<Operation> operations) {
+        this.operations = operations;
+    }
+}