+++ data-transfer-object/src/main/java/org/grnet/status/dtos/profile/operation/TruthTableEntry.java
@@ -1,0 +1,33 @@
+package org.grnet.status.dtos.profile.operation;
+
+public class TruthTableEntry {
+
+    private String a;
+    private String b;
+    private String x;
+
+    public String getA() {
+        return a;
+    }
+
+    public void setA(String a) {
+        this.a = a;
+    }
+
+    public String getB() {
+        return b;
+    }
+
+    public void setB(String b) {
+        this.b = b;
+    }
+
+    public String getX() {
+        return x;
+    }
+
+    public void setX(String x) {
+        this.x = x;
+    }
+}
+