+++ service/src/main/java/org/grnet/status/services/clients/ArgoWebApiClient.java
@@ -5,8 +5,10 @@ import jakarta.ws.rs.core.MediaType;
 import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
 import org.grnet.status.dtos.argo.ArgoReportsResponse;
 import org.grnet.status.dtos.argo.ArgoStatusGroupsResponse;
+import org.grnet.status.dtos.profile.aggregation.AggregationProfileResponse;
+import org.grnet.status.dtos.profile.metric.MetricProfileResponse;
+import org.grnet.status.dtos.profile.operation.OperationProfileResponse;
 import org.grnet.status.dtos.report.WebApiReportResponse;
-import org.grnet.status.dtos.tenant.TenantRequestDto;
 import org.grnet.status.dtos.tenant.webapi.TenantWebApiGetResponse;
 import org.grnet.status.dtos.tenant.webapi.TenantWebApiCreateResponse;
 import org.grnet.status.dtos.tenant.webapi.TenantWebApiRequest;
@@ -78,12 +80,39 @@ public interface ArgoWebApiClient {
     @PUT
     @Path("/api/v2/admin/tenants/{id}/db-conf")
     TenantWebApiResponse updateTenantDBConf( @PathParam("id") String id,
-                                              @HeaderParam("x-api-key") String apiKey,
-                                              TenantWebApiRequest request
-    ) throws WebApplicationException, ProcessingException;
+                                              @HeaderParam("x-api-key") String apiKey, TenantWebApiRequest request) throws WebApplicationException, ProcessingException;
+
+    @GET
+    @Path("/api/v2/operations_profiles/{id}")
+    OperationProfileResponse listSpecificOperationsProfiles(@PathParam("id") String id, @QueryParam("date") String date,
+                                                            @HeaderParam("x-api-key") String apiKey) throws WebApplicationException, ProcessingException;
 
     @GET
     @Path("/api/v2/reports/{id}")
     WebApiReportResponse fetchReportById(@PathParam("id") String id, @HeaderParam("x-api-key") String apiKey) throws WebApplicationException, ProcessingException;
 
+    @GET
+    @Path("/api/v2/operations_profiles")
+    OperationProfileResponse listAllOperationsProfiles(@QueryParam("date") String date,
+                                                            @HeaderParam("x-api-key") String apiKey) throws WebApplicationException, ProcessingException;
+
+    @GET
+    @Path("/api/v2/aggregation_profiles/{id}")
+    AggregationProfileResponse listSpecificAggregationProfiles(@PathParam("id") String id, @QueryParam("date") String date,
+                                                               @HeaderParam("x-api-key") String apiKey) throws WebApplicationException, ProcessingException;
+
+    @GET
+    @Path("/api/v2/aggregation_profiles")
+    AggregationProfileResponse listAllAggregationProfiles(@QueryParam("date") String date,
+                                                       @HeaderParam("x-api-key") String apiKey) throws WebApplicationException, ProcessingException;
+
+    @GET
+    @Path("/api/v2/metric_profiles/{id}")
+    MetricProfileResponse listSpecificMetricProfiles(@PathParam("id") String id, @QueryParam("date") String date,
+                                                     @HeaderParam("x-api-key") String apiKey) throws WebApplicationException, ProcessingException;
+
+    @GET
+    @Path("/api/v2/metric_profiles")
+    MetricProfileResponse listAllMetricProfiles(@QueryParam("date") String date,
+                                                          @HeaderParam("x-api-key") String apiKey) throws WebApplicationException, ProcessingException;
 }